2. Component. ADDED styles to pages
