+++ src/main.jsx
@@ -2,10 +2,10 @@ import ReactDOM from 'react-dom/client'
 import {createBrowserRouter, RouterProvider} from 'react-router-dom'
 import App from './App'
 import './index.css'
-import AboutMe from './pages/about'
-import Contact from './pages/contact'
-import ProjectList from './pages/portfolio'
-import Resume from './pages/resume'
+import AboutMe from './pages/about/about'
+import Contact from './pages/contact/contact'
+import ProjectList from './pages/portfolio/portfolio'
+import Resume from './pages/resume/resume'
 
 const router = createBrowserRouter([
   {
