+++ src/pages/portfolio/portfolio.css
@@ -1,0 +1,48 @@
+.project-container{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 3rem;
+    margin-top:5rem;
+    margin-bottom: 5rem;
+    
+}
+.project-card{
+    display: flex;
+    flex-direction: column;
+    background-color: rgba(240, 248, 255, 0.15);
+    max-width: 50rem;
+}
+
+.project-title{
+    font-size:x-large;
+    margin: 1rem;
+}
+
+.project-image{
+    padding: 1rem;
+}
+
+.project-link-container{
+    display: flex;
+    gap: 1rem;
+    justify-content: center;
+    margin-bottom: 1rem;
+}
+
+.link-container{
+    background-color: rgba(0, 0, 0, 0.25);
+    padding: 0.5rem;
+    min-width: 3rem;
+}
+
+/* .project-link{
+    transition: transform 0.3s;
+}
+
+.project-link:hover{
+    transform: scale(1.5);
+    transition: 0.3s;
+} */
+
+/* doesn't seem to work for some reason */
